requirements.txt
langchain
openai
python-dotenv
wikipedia
duckduckgo-search